requirements.txt
fastrtc[vad]>=0.0.19
groq>=0.22.0
numpy>=2.1.3
langgraph>=0.1.18
langchain-core>=0.1.29
langchain-groq>=0.1.5
langchain-openai>=0.1.0
langchain-community>=0.0.20
loguru>=0.7.3
python-dotenv>=1.0.0
deepgram-sdk>=4.0.0
pyaudio>=0.2.14
faiss-cpu>=1.7.4
pypdf>=4.0.0
rapidocr-onnxruntime>=1.3.22
pymupdf>=1.23.0
langchain-openai>=0.1.0
langchain-community>=0.0.20
langchain-text-splitters